requirements.txt
fastapi
sqlalchemy
uvicorn
pydantic
psycopg2-binary
requests
streamlit
python-dotenv
types-requests
jose
types-python-jose